Implement database (closes #16)

+++ src/main/java/shortener/database/entities/UserSession.java
@@ -6,5 +6,5 @@ import com.fasterxml.jackson.annotation.JsonAutoDetect;
  * UserSession record type.
  */
 @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
-public record UserSession(Integer userId, String token) {
+public record UserSession(Long userId, String token) {
 }
